Load tags dynamically using query string
Instead of having a separate .html file for each tag, now there's a single tag.html that takes in a query string of the tag and displays all posts under that tag. This change also makes the separate "tags" folder useless so I moved the tag-related files into the main directory, and renamed the index's script and style files to homeScript and homeStyle respectively.